+++ rn_attachment_preview/tests/__init__.py
@@ -2,3 +2,4 @@
 
 from . import test_attachment_preview
 from . import test_attachment_preview_capture
+from . import test_attachment_preview_v2